[Port] Core 31 compile-and-test fixes for the Freicoin delta
Compiler/test-guided pass after transplanting the series onto v31.1.

Structural reconstructions (v31 rewrote these subsystems; a hunk-level
merge produced non-compiling chimeras, so each was rebuilt as pristine
v31.1 + the v30.0..rebase-30 Freicoin delta re-applied):
- script/miniscript.{h,cpp}: v31 moved Node to value semantics
  (NodeRef/MakeNodeRef gone, subs->Subs() accessors); re-applied the
  taproot removal (MULTI_A, TAPSCRIPT context, x-only keys) and
  Freicoin's VERIFY-typed OLDER/AFTER + MultiSigHint satisfier.
- node/miner.{h,cpp}: v31 replaced ancestor-score selection with
  cluster linearization; kept v31 selection verbatim and re-grafted the
  block-final state machine, vTxFees[0] placeholder, coinbase MTP
  locktime with SEQUENCE_FINAL, aux-pow template setup, and the
  waitNext template-fee accounting.
- script/sign.{h,cpp}, signingprovider, scriptpubkeyman, descriptor:
  excised MuSig2 machinery (taproot-only) that v31 threaded through the
  signing stack.

Consensus-relevant fix caught by the reorg unit tests:
- CCoinsViewCache::Reset() (new in v31's per-connect reset guard) must
  also clear the cached block-final entry; otherwise after a reorg
  across a block-final boundary the stale entry survives and ConnectBlock
  fatals with "prior block-final tx hash not found; corruption likely".
  On a live node any such reorg would have halted the daemon.
- consensus/merkle.cpp MerkleComputation: restored the *proot/*pmutated
  output writes for the empty-leaves and final-result paths (dropped in
  the merge); without them Freicoin's block-final/aux-pow commitment
  hashes were wrong, diverging the test-chain tip.

Mechanical v31 adaptations: script_verify_flags strong type (the flag
locals, VerifyWitnessProgram/CheckInputScripts/CScriptCheck signatures,
TrimFlags, AllConsensusFlags); GetScriptFlagNames ported with the
Freicoin flag set; logging categories.h + STRATUM; hashBlock back to
plain uint256 to match v31's CoinsViewCache; dbwrapper GetObfuscation;
FeeRateFormat::FRC_KVB; Parse*->ToIntegral shims; Txid/Wtxid .ToUint256();
GetBlockProof/GetBitsProof moved to chain.cpp with aux-pow work; Coin and
CNode test ctors; descriptor key_exp_index increment for bare privkeys.

Unit suite: tx_valid/tx_invalid/script_json (the data-driven consensus
script tests) and the full reorg/assumeutxo/wallet suites pass. Known
non-consensus test-harness residue (see doc/REBASE-NOTES.md): script_build
JSON-regen mismatch on one uncompressed-P2WSH MAST witness vector;
util_tests/test_LockDirectory and coinselector bnb are green in isolation
but flaky under the full run's forked/global state on v31.

+++ src/test/transaction_tests.cpp
@@ -169,8 +169,8 @@ bool CheckTxScripts(const CTransaction& tx, const std::map<COutPoint, CScript>& 
 script_verify_flags TrimFlags(script_verify_flags flags)
 {
     // Clear verification flags related to protocl rule changes
-    flags &= ~(unsigned int)SCRIPT_VERIFY_PROTOCOL_CLEANUP;
-    flags &= ~(unsigned int)SCRIPT_VERIFY_SIZE_EXPANSION;
+    flags &= ~SCRIPT_VERIFY_PROTOCOL_CLEANUP;
+    flags &= ~SCRIPT_VERIFY_SIZE_EXPANSION;
 
     // WITNESS requires P2SH
     if (!(flags & SCRIPT_VERIFY_P2SH)) flags &= ~SCRIPT_VERIFY_WITNESS;
